Nice optimizations for UserIdGenerator that signficantly cuts down on the processing overhead. Now we're going to look into fixing up the AuctionMarkProfile.
git-svn-id: https://oltpbenchmark.googlecode.com/svn/trunk@341 8c9255d6-0b70-036a-fcb6-8a999947b34c

+++ tests/com/oltpbenchmark/benchmarks/auctionmark/util/TestUserIdGenerator.java
@@ -76,14 +76,14 @@ public class TestUserIdGenerator extends TestCase {
 	 */
 	public void testSeekToPosition() throws Exception {
 	    UserIdGenerator generator = new UserIdGenerator(users_per_item_count, 1);
-	    final long num_users = generator.getTotalUsers()-1;
+	    final int num_users = (int)(generator.getTotalUsers()-1);
 	    
 	    int itemCount = rand.nextInt(users_per_item_count.getMaxValue().intValue()-1);
 	    generator.setCurrentItemCount(itemCount);
 	    System.err.println("itemCount = " + itemCount);
 	    
-	    long cur_position = generator.getCurrentPosition();
-        long new_position = rand.number(cur_position, num_users);
+	    int cur_position = generator.getCurrentPosition();
+        int new_position = rand.number(cur_position, num_users);
 //        System.err.println(users_per_item_count);
         System.err.println("cur_position = " + cur_position);
         System.err.println("new_position = " + new_position);
